End of project

+++ controllers/contact_controller.js
@@ -6,7 +6,7 @@ const Contact= require("../models/contactModel");
 
 
 const getContacts =asyncHandler(async(req,res) => {
-    const contacts=await Contact.find();
+    const contacts=await Contact.find({user_id:req.user.id});
     res.status(200).json(contacts);
 });
 
@@ -20,7 +20,8 @@ const createContact =asyncHandler(async(req,res) => {
     const contact=await Contact.create({
         name,
         email,
-        phone
+        phone,
+        user_id:req.user.id
     });
     res.status(200).json({message:"create Contacts"});
 });
@@ -31,12 +32,17 @@ const updateContact =asyncHandler(async(req,res) => {
         res.status(404);
         throw new Error("Contact does not exist");
     }
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User Don't have permission to access the contact!");
+    }
     const updatedContact= await Contact.findByIdAndUpdate(
         req.params.id,
         req.body,
         {new:true}
 
 );
+        
     res.status(200).json(updatedContact);
 });
 
@@ -56,6 +62,10 @@ const deleteContact =asyncHandler(async(req,res) => {
     if(!contact){
         res.status(404);
         throw new Error("Contact does not exist");
+    }
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User Don't have permission to access the contact!");
     }
     await Contact.findByIdAndRemove(req.params.id);
     res.status(200).json(contact);
